+++ pkg/cmd/openshift-operators/webconsole-operator/controller/webconsole_operator_trigger.go
@@ -4,13 +4,9 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/golang/glog"
-
-	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	utilerrors "k8s.io/apimachinery/pkg/util/errors"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/apimachinery/pkg/util/wait"
 	appsclientv1 "k8s.io/client-go/kubernetes/typed/apps/v1"
@@ -19,6 +15,7 @@ import (
 	apiregistrationclientv1beta1 "k8s.io/kube-aggregator/pkg/client/clientset_generated/clientset/typed/apiregistration/v1beta1"
 	"k8s.io/kubernetes/pkg/api/legacyscheme"
 
+	"github.com/golang/glog"
 	webconsoleconfigv1 "github.com/openshift/api/webconsole/v1"
 	"github.com/openshift/origin/pkg/cmd/openshift-operators/util/resourceapply"
 	"github.com/openshift/origin/pkg/cmd/openshift-operators/util/resourcemerge"
@@ -83,15 +80,13 @@ func (c WebConsoleOperator) sync() error {
 	// TODO use semver
 	isFirst := len(operatorConfig.Status.LastSuccessfulVersion) == 0
 	isSame := operatorConfig.Spec.Version == operatorConfig.Status.LastSuccessfulVersion
-	is10_0 := operatorConfig.Status.LastSuccessfulVersion == "3.10.0" || operatorConfig.Status.LastSuccessfulVersion == "3.10"
-	wants10_0 := operatorConfig.Spec.Version == "3.10.0" || operatorConfig.Spec.Version == "3.10"
+	is10_0 := is10_0Version(operatorConfig.Status.LastSuccessfulVersion)
+	wants10_0 := is10_0Version(operatorConfig.Spec.Version)
 	wants10_1 := operatorConfig.Spec.Version == "3.10.1"
-
-	fmt.Printf("#### %v %v %v %v %v \n", isFirst, isSame, is10_0, wants10_0, wants10_1)
 
 	switch {
 	case wants10_0 && (isSame || isFirst):
-		return c.sync10(operatorConfig)
+		return c.sync10_0(operatorConfig)
 
 	case wants10_1 && (isSame || isFirst):
 		return c.sync10_1(operatorConfig)
@@ -103,201 +98,6 @@ func (c WebConsoleOperator) sync() error {
 		// TODO update status
 		return fmt.Errorf("unrecognized state")
 	}
-}
-
-// between 3.10.0 and 3.10.1, we want to fix some naming mistakes.  During this time we need to ensure the
-// 3.10.0 resources are correct and we need to spin up 3.10.1 *before* removing the old
-func (c WebConsoleOperator) migrate10_0_to_10_1(operatorConfig *webconsolev1.OpenShiftWebConsoleConfig) error {
-	errors := []error{}
-	fmt.Printf("#### 1a\n")
-	if _, err := c.ensureNamespace(); err != nil {
-		errors = append(errors, err)
-	}
-
-	fmt.Printf("#### 1b\n")
-	if _, err := resourceapply.ApplyServiceAccount(c.corev1Client,
-		&corev1.ServiceAccount{ObjectMeta: metav1.ObjectMeta{Namespace: "openshift-web-console", Name: "web-console"}}); err != nil {
-		errors = append(errors, err)
-	}
-
-	// keep 3.10.0 up to date in case we have to do things like change the config
-	fmt.Printf("#### 1c\n")
-	if err := c.sync10(operatorConfig); err != nil {
-		errors = append(errors, err)
-	}
-
-	// create the 3.10.1 resources
-	fmt.Printf("#### 1d\n")
-	modifiedConfig, err := c.ensureWebConsoleConfigMap10_1(operatorConfig.Spec)
-	if err != nil {
-		errors = append(errors, err)
-	}
-	fmt.Printf("#### 1e\n")
-	modifiedDeployment, err := c.ensureWebConsoleDeployment10_1(operatorConfig.Spec)
-	if err != nil {
-		errors = append(errors, err)
-	}
-
-	// if we modified the 3.10.1 resources, then we have to wait for them to resettle.  Return and we'll be requeued on the watch
-	fmt.Printf("#### 1f\n")
-	if modifiedConfig || modifiedDeployment {
-		// TODO update intermediate status and return
-		return utilerrors.NewAggregate(errors)
-	}
-	fmt.Printf("#### 1g\n")
-	if len(errors) > 0 {
-		return utilerrors.NewAggregate(errors)
-	}
-
-	// TODO check to see if the deployment has ready pods
-	fmt.Printf("#### 1h\n")
-	newDeployment, err := c.appsv1Client.Deployments("openshift-web-console").Get("web-console", metav1.GetOptions{})
-	if err != nil {
-		// TODO update intermediate status
-		return err
-	}
-	fmt.Printf("#### 1i\n")
-	deployment10_1IsReady := newDeployment.Status.ReadyReplicas > 0
-	if !deployment10_1IsReady {
-		// TODO update intermediate status
-		return utilerrors.NewAggregate(errors)
-	}
-
-	// after the new deployment has ready pods, we can swap the service and remove the old resources
-	fmt.Printf("#### 1j\n")
-	if _, err := resourceapply.ApplyService(c.corev1Client, resourceread.ReadServiceOrDie([]byte(service10_1Yaml))); err != nil {
-		errors = append(errors, err)
-	}
-
-	// if we have errors at this point, we cannot remove the 3.10.0 resources
-	fmt.Printf("#### 1k\n")
-	if len(errors) > 0 {
-		// TODO update status
-		return utilerrors.NewAggregate(errors)
-	}
-
-	fmt.Printf("#### 1l\n")
-	if err := c.appsv1Client.Deployments("openshift-web-console").Delete("webconsole", nil); err != nil {
-		errors = append(errors, err)
-	}
-
-	fmt.Printf("#### 1m\n")
-	if err := c.corev1Client.ConfigMaps("openshift-web-console").Delete("webconsole-config", nil); err != nil {
-		errors = append(errors, err)
-	}
-
-	// set the status
-	fmt.Printf("#### 1n\n")
-	operatorConfig.Status.LastUnsuccessfulRunErrors = []string{}
-	for _, err := range errors {
-		operatorConfig.Status.LastUnsuccessfulRunErrors = append(operatorConfig.Status.LastUnsuccessfulRunErrors, err.Error())
-	}
-	if len(errors) == 0 {
-		operatorConfig.Status.LastSuccessfulVersion = operatorConfig.Spec.Version
-	}
-	if _, err := c.operatorConfigClient.OpenShiftWebConsoleConfigs().Update(operatorConfig); err != nil {
-		// if we had no other errors, then return this error so we can re-apply and then re-set the status
-		if len(errors) == 0 {
-			return err
-		}
-		utilruntime.HandleError(err)
-	}
-
-	return utilerrors.NewAggregate(errors)
-}
-
-func (c WebConsoleOperator) sync10_1(operatorConfig *webconsolev1.OpenShiftWebConsoleConfig) error {
-	errors := []error{}
-	// TODO the configmap and secret changes for daemonset should actually be a newly created configmap and then a subsequent daemonset update
-	// TODO this requires us to be able to detect that the changes have not worked well and trigger an effective rollback to previous config
-	if _, err := c.ensureNamespace(); err != nil {
-		errors = append(errors, err)
-	}
-
-	if _, err := resourceapply.ApplyService(c.corev1Client, resourceread.ReadServiceOrDie([]byte(service10_1Yaml))); err != nil {
-		errors = append(errors, err)
-	}
-
-	if _, err := resourceapply.ApplyServiceAccount(c.corev1Client,
-		&corev1.ServiceAccount{ObjectMeta: metav1.ObjectMeta{Namespace: "openshift-web-console", Name: "web-console"}}); err != nil {
-		errors = append(errors, err)
-	}
-
-	// we need to make a configmap here
-	if _, err := c.ensureWebConsoleConfigMap10_1(operatorConfig.Spec); err != nil {
-		errors = append(errors, err)
-	}
-
-	// our configmaps and secrets are in order, now it is time to create the DS
-	// TODO check basic preconditions here
-	if _, err := c.ensureWebConsoleDeployment10_1(operatorConfig.Spec); err != nil {
-		errors = append(errors, err)
-	}
-
-	// set the status
-	operatorConfig.Status.LastUnsuccessfulRunErrors = []string{}
-	for _, err := range errors {
-		operatorConfig.Status.LastUnsuccessfulRunErrors = append(operatorConfig.Status.LastUnsuccessfulRunErrors, err.Error())
-	}
-	if len(errors) == 0 {
-		operatorConfig.Status.LastSuccessfulVersion = operatorConfig.Spec.Version
-	}
-	if _, err := c.operatorConfigClient.OpenShiftWebConsoleConfigs().Update(operatorConfig); err != nil {
-		// if we had no other errors, then return this error so we can re-apply and then re-set the status
-		if len(errors) == 0 {
-			return err
-		}
-		utilruntime.HandleError(err)
-	}
-
-	return utilerrors.NewAggregate(errors)
-}
-
-func (c WebConsoleOperator) sync10(operatorConfig *webconsolev1.OpenShiftWebConsoleConfig) error {
-	errors := []error{}
-	// TODO the configmap and secret changes for daemonset should actually be a newly created configmap and then a subsequent daemonset update
-	// TODO this requires us to be able to detect that the changes have not worked well and trigger an effective rollback to previous config
-	if _, err := c.ensureNamespace(); err != nil {
-		errors = append(errors, err)
-	}
-
-	if _, err := resourceapply.ApplyService(c.corev1Client, resourceread.ReadServiceOrDie([]byte(service10Yaml))); err != nil {
-		errors = append(errors, err)
-	}
-
-	if _, err := resourceapply.ApplyServiceAccount(c.corev1Client,
-		&corev1.ServiceAccount{ObjectMeta: metav1.ObjectMeta{Namespace: "openshift-web-console", Name: "webconsole"}}); err != nil {
-		errors = append(errors, err)
-	}
-
-	// we need to make a configmap here
-	if _, err := c.ensureWebConsoleConfigMap10(operatorConfig.Spec); err != nil {
-		errors = append(errors, err)
-	}
-
-	// our configmaps and secrets are in order, now it is time to create the DS
-	// TODO check basic preconditions here
-	if _, err := c.ensureWebConsoleDeployment10(operatorConfig.Spec); err != nil {
-		errors = append(errors, err)
-	}
-
-	// set the status
-	operatorConfig.Status.LastUnsuccessfulRunErrors = []string{}
-	for _, err := range errors {
-		operatorConfig.Status.LastUnsuccessfulRunErrors = append(operatorConfig.Status.LastUnsuccessfulRunErrors, err.Error())
-	}
-	if len(errors) == 0 {
-		operatorConfig.Status.LastSuccessfulVersion = operatorConfig.Spec.Version
-	}
-	if _, err := c.operatorConfigClient.OpenShiftWebConsoleConfigs().Update(operatorConfig); err != nil {
-		// if we had no other errors, then return this error so we can re-apply and then re-set the status
-		if len(errors) == 0 {
-			return err
-		}
-		utilruntime.HandleError(err)
-	}
-
-	return utilerrors.NewAggregate(errors)
 }
 
 func (c WebConsoleOperator) ensureNamespace() (bool, error) {
@@ -328,58 +128,6 @@ func readWebConsoleConfiguration(objBytes string) (*webconsoleconfigv1.WebConsol
 	}
 
 	return ret, nil
-}
-
-func (c WebConsoleOperator) ensureWebConsoleConfigMap10(options webconsolev1.OpenShiftWebConsoleConfigSpec) (bool, error) {
-	requiredConfig, err := ensureWebConsoleConfig(options)
-	if err != nil {
-		return false, err
-	}
-
-	newWebConsoleConfig, err := runtime.Encode(legacyscheme.Codecs.LegacyCodec(webconsoleconfigv1.SchemeGroupVersion), requiredConfig)
-	if err != nil {
-		return false, err
-	}
-	requiredConfigMap := resourceread.ReadConfigMapOrDie([]byte(configMap10Yaml))
-	requiredConfigMap.Data[configConfigMap10Key] = string(newWebConsoleConfig)
-
-	return resourceapply.ApplyConfigMap(c.corev1Client, requiredConfigMap)
-}
-
-func (c WebConsoleOperator) ensureWebConsoleDeployment10(options webconsolev1.OpenShiftWebConsoleConfigSpec) (bool, error) {
-	required := resourceread.ReadDeploymentOrDie([]byte(deployment10Yaml))
-	required.Spec.Template.Spec.Containers[0].Image = options.ImagePullSpec
-	required.Spec.Template.Spec.Containers[0].Args = append(required.Spec.Template.Spec.Containers[0].Args, fmt.Sprintf("-v=%d", options.LogLevel))
-	required.Spec.Replicas = &options.Replicas
-	required.Spec.Template.Spec.NodeSelector = options.NodeSelector
-
-	return resourceapply.ApplyDeployment(c.appsv1Client, required)
-}
-
-func (c WebConsoleOperator) ensureWebConsoleConfigMap10_1(options webconsolev1.OpenShiftWebConsoleConfigSpec) (bool, error) {
-	requiredConfig, err := ensureWebConsoleConfig(options)
-	if err != nil {
-		return false, err
-	}
-
-	newWebConsoleConfig, err := runtime.Encode(legacyscheme.Codecs.LegacyCodec(webconsoleconfigv1.SchemeGroupVersion), requiredConfig)
-	if err != nil {
-		return false, err
-	}
-	requiredConfigMap := resourceread.ReadConfigMapOrDie([]byte(configMap10_1Yaml))
-	requiredConfigMap.Data[configConfigMap10_1Key] = string(newWebConsoleConfig)
-
-	return resourceapply.ApplyConfigMap(c.corev1Client, requiredConfigMap)
-}
-
-func (c WebConsoleOperator) ensureWebConsoleDeployment10_1(options webconsolev1.OpenShiftWebConsoleConfigSpec) (bool, error) {
-	required := resourceread.ReadDeploymentOrDie([]byte(deployment10_1Yaml))
-	required.Spec.Template.Spec.Containers[0].Image = options.ImagePullSpec
-	required.Spec.Template.Spec.Containers[0].Args = append(required.Spec.Template.Spec.Containers[0].Args, fmt.Sprintf("-v=%d", options.LogLevel))
-	required.Spec.Replicas = &options.Replicas
-	required.Spec.Template.Spec.NodeSelector = options.NodeSelector
-
-	return resourceapply.ApplyDeployment(c.appsv1Client, required)
 }
 
 // Run starts the webconsole and blocks until stopCh is closed.
